Implemented spaces in the Engine
Spaces are now used.  An active space can bet set and it will be used
for all updates automatically. Whenever the active space is changed,
that space will be used for update.

A default space, "Game World" is now created in Init() by the engine
automatically and set to be the default active space.