Added Archetype Factory
The engine now has an Archetype Factory.

Archetypes allow easy creation of generic object types.

An archetype can be created using the format of ArchetypeCreateFn, and
then added to the factory by naming the Archetype (IE "Default Camera")
and passing the pointer to the creation function to the Factory
addArchetype function. The addArchetype function calls should be put in
the ArchetypeFactory constructor.