typedef pointers, GETSYS, Camera system
GETSYS macro can be used to get any system that is currently active in
the engine by passing in the typename of the system.

The camera system is used to update the camera matrices based on any
changes that occurred with the camera's position, scale, etc. The
matrices are then used by the GLGraphics system to draw objects
correctly.

Some classes (components, systems, etc) didn't have typedef pointers,
now they do.