Spaces now use the same syntax as everything else
Previously spaces were accessed from the engine via reference, which
made for awkward changes in syntax when doing operations from the engine
on spaces, ie:
ENGINE->GetActiveSpace().GetCamera()->GET_COMPONENT(Camera)->_nearPlane;

It was confusing and unnecessary, and so spaces are now accessed via
smartpointer using SpacePtr objects much like everything else in the
engine. The new syntax for the previous operation would be as follows:
ENGINE->GetActiveSpace()->GetCamera()->GET_COMPONENT(Camera)->_nearPlane

Note that in the first example, the GetActiveSpace() call had a '.'
operator, while everything else was using '->' meanwhile in the second
example, the syntax is consistent and they all use the '->'

This results in simpler syntax.