Gamestate Management Infrastructure
Moved the ENGINE pointer assignment into the constructor of Engine and
out of main().

Implemented the interface for a Gamestate. Gamestates allow easy menu
navigation implementation, and an easy way to choose which spaces are
active.

The Engine interfaces to use states have been hooked up and implemented,
so it is possible to push and pop gamestates, as well as get access to
the current gamestate.

The Engine currently does not utilize the gamestates, that functionality
will be included in my next update.

The newFrame() call that previously resided in GLGraphics::Update(float)
has been moved to the engine's Update(float). This is so that multiple
spaces can draw in a single frame, in order to allow space usage for
features such as a HUD or background effect. That enables the objects to
exist in separate spaces and still be shown together every frame.

TODO: Convert TestSystem into a gamestate and implement a basic main
menu and pause menu to showcase the usage.