Updating now uses gamestates
Gamestates are now implemented and in use.

Gamestates can register and deregister which spaces should be updated,
and they can set which space will be used for the gamestate's Update()
logic.

I also did some small organizational changes to go along with converting
testsystem into a gamestate. There is now a Gamestates folder, complete
with a GamestatesInclude.h.

Current build verified working in debug & release of both 32 and 64 bit
builds.

TODO: Stack gamestate changes and apply them at the end of the frame.
This may not be necessary, requires further investigation.