Added menus and menu navigation
This was a rather large update so bear with me here:

Added a menuObject component. It has a menutype (currently MAIN or
PAUSE) and a menu action (there are various options for pause and main
menu actions, 9 total). I have added an option for each of them to the
archetype factory so they can all be created easily.

Pressing escape now sends a message instead of immediately terminating.
It is now used to open the pause menu from gameplay, closing the pause
menu while in it, and exiting the program from the main menu.

Point collision can now be checked by the physics system. It
automatically checks collision with the mouse against objects now and
sends a hover or click message whenever an object is colliding with the
mouse, depending on whether the left mouse button is pressed or not.

The menu buttons make use of the engine's messaging as well, by
receiving the messages sent when a mouse click message is sent. The
message handler in the pause and main menu states will determine the
type of button being clicked and respond accordingly.

Currently the only buttons actually hooked up with logic are the main
menu start game and exit buttons, and the pause menu resume game and
quit game buttons. These are the green and red buttons. The yellow
buttons for how to play, credits, and options have not been hooked up as
there are no gamestates for those buttons yet.