Added buttonController, fixed input lag.
Updated the menu button archetypes to include setting their color.

The buttons now have a controller class that causes them to change color
when hovered. This is used to produce a nice brightening effect when the
user has moused over them.

Also switched the input checking function to a loop so that it will
parse all input messages instead of just the oldest one. This has ended
the input lag caused by pressing multiple keys while moving the mouse.