Implemented several gameplay components and archetypes
Bullet - stores a damage value
Health - stores current and max HP
Score - can be used to track an objective's score value or a units
current score.
Weapon - Has a recharge rate and timer, and some values relating to shot
tracking.

Implemented archetypes for a player, a players bullet, and an asteroid.