lifetime component, bulletController
Components.h has been renamed to ComponentsInclude.h to match the naming
convention used for the other include headers.

Lifetime component can be used for things that have a limited lifetime,
such as a bullet.

The bullet controller handles the lifetime of bullets and resolving
collisions that involve bullets. This includes score modification for
the shooter, recording the hit/miss, and destruction of the bullet at
the end of its life.